dom/camera/test/test_camera_hardware_auto_focus_moving_cb.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=965421
-->
<head>
  <title>Bug 965421 - Test camera hardware API for Auto focus moving Callback</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="camera_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=965421">Mozilla Bug 965421</a>
  <video id="viewfinder" width = "200" height = "200" autoplay></video>
  <img src="#" alt="This image is going to load" id="testimage"/>

<script class="testbody" type="text/javascript;version=1.7">

var suite = new CameraTestSuite();

suite.test('auto-focus-moving', function() {
  function triggerAutoFocusMoving(p) {
    var sync = new Promise(function(resolve, reject) {
      function onEvent(e) {
        suite.camera.removeEventListener('focus', onEvent);
        ok(e.newState === 'focusing', 'autofocus event state focusing == ' + e.newState);
        resolve();
      }
      suite.camera.addEventListener('focus', onEvent);
    });

    suite.hw.fireAutoFocusMoving(true);
    return sync;
  }

  function waitAutoFocusComplete(p) {
    var sync = new Promise(function(resolve, reject) {
      function onEvent(e) {
        suite.camera.removeEventListener('focus', onEvent);
        ok(e.newState === 'focused', 'autofocus event state focused == ' + e.newState);
        resolve();
      }
      suite.camera.addEventListener('focus', onEvent);
    });

    // Missing the fireAutoFocusComplete but it should timeout on its own
    suite.hw.fireAutoFocusMoving(false);
    return sync;
  }

  function runAutoFocusCycle(p) {
    return triggerAutoFocusMoving(p)
      .then(waitAutoFocusComplete);
  }

  /* If the driver doesn't supply an onAutoFocusComplete notification,
     gecko will timeout and provide it. After three times, it will no
     longer rely upon a timeout and fire it immediately. */
  return suite.getCamera()
    .then(runAutoFocusCycle)
    .then(runAutoFocusCycle)
    .then(runAutoFocusCycle)
    .then(runAutoFocusCycle);
});

suite.setup()
  .then(suite.run);

</script>
</body>

</html>
